fixing warnings
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@27012 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ ftp-casing/localapi/com/untangle/node/ftp/FtpUtil.java
@@ -119,7 +119,7 @@ class FtpUtil
         StringBuffer sb = new StringBuffer();
         byte[] addr = socketAddress.getAddress().getAddress();
         for (byte a : addr) {
-            sb.append(((int)a) & 0xFF);
+            sb.append((a) & 0xFF);
             sb.append(',');
         }
